New IDE

+++ libs/Trafficator/Trafficator.h
@@ -4,21 +4,20 @@
 #include <Settings/Settings.h>
 #include "Arduino.h"
 
-class Trafficator
-{
+class Trafficator {
 private:
-        int pinout;
-        bool isEnabled;
-        int ledState;
-        long previousMillis;
-        long interval;
+	uint8_t pinout;
+	bool isEnabled;
+	int ledState;
+	long previousMillis;
+	long interval;
 
 public:
-        Trafficator(int pin);
-        void update();
-        void setup();
-        void on();
-        void off();
+	Trafficator(uint8_t pin);
+	void update();
+	void setup();
+	void on();
+	void off();
 };
 
 #endif
